+++ cdk-examples/rust-standalone/lib/rust-standalone-stack.ts
@@ -2,9 +2,9 @@ import { Duration, Stack, StackProps } from 'aws-cdk-lib';
 import * as s3 from 'aws-cdk-lib/aws-s3';
 import { BlockPublicAccess } from 'aws-cdk-lib/aws-s3';
 import { Construct } from 'constructs';
-import { RustFunction } from 'rust.aws-cdk-lambda';
+// import { RustFunction } from 'rust.aws-cdk-lambda';
 // uncomment for local testing
-// import { RustFunction } from '../../../lib';
+import { RustFunction } from '../../../lib';
 
 export class RustStandaloneStack extends Stack {
     constructor(scope: Construct, id: string, props?: StackProps) {
@@ -25,6 +25,11 @@ export class RustStandaloneStack extends Stack {
             },
             // Useful so library logs show up in CloudWatch
             setupLogging: true,
+            // Enable optional features and env variables at build (compile) time.
+            features: ['my-second-feature'],
+            buildEnvironment: {
+                MY_BUILD_ENV_VAR: 'Testing 123.',
+            },
         });
 
         bucket.grantReadWrite(myLambda);